+++ CHANGELOG.md
@@ -11,12 +11,14 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - elastic_class.ElasticSearch: Added get_shards method.
 - elastic_class.ElasticSearch: Added get_repo_list method.
 - elastic_class.ElasticSearch: Added get_nodes method.
+- elastic_class.ElasticSearch: Added get_master_name method.
 
 ### Deprecated
 - elastic_class.is_active function.
 - elastic_class.get_shards function.
 - elastic_class.get_repo_list function.
 - elastic_class.get_nodes function.
+- elastic_class.get_master_name function.
 - elastic.py: The port and scheme entries in the configuration file.